pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=61.0", "setuptools_scm>=8"]

[tool.setuptools_scm]

[project]
name = "bambi"
description = "BAyesian Model Building Interface in Python"
requires-python = ">=3.10,<3.13"
readme = "README.md"
license = {file = "LICENSE"}
dynamic = ["version"]
maintainers = [
    { name = "Tomás Capretto", email = "tomicapretto@gmail.com" },
    { name = "Osvaldo Martin", email = "aloctavodia@gmail.com" },
    { name = "Gabriel Stechschulte", email = "stechschulteg@gmail.com" },
    { name = "Ravin Kumar", email = "ravinsdrive@gmail.com" },
]

dependencies = [
    "arviz>=0.12.0",
    "formulae>=0.5.3",
    "graphviz",
    "pandas>=1.0.0",
    "pymc>=5.15.1",
]

[project.optional-dependencies]
dev = [
    "black==24.3.0",
    "ipython>=5.8.0,!=8.7.0",
    "pre-commit>=2.19",
    "pylint==3.1.0",
    "pytest-cov>=2.6.1",
    "pytest>=4.4.0",
    "quartodoc==0.6.1",
    "seaborn>=0.9.0",
]

jax = [
    "bayeux-ml>=0.1.12",
]

[project.urls]
homepage = "https://bambinos.github.io/bambi"
documentation = "https://bambinos.github.io/bambi"
repository = "https://github.com/bambinos/bambi"
changelog = "https://github.com/bambinos/bambi/blob/main/docs/CHANGELOG.md"

[tool.setuptools]
packages = [
    "bambi", 
    "bambi.backend", 
    "bambi.data", 
    "bambi.defaults", 
    "bambi.families",
    "bambi.interpret", 
    "bambi.priors", 
    "bambi.terms", 
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
addopts = "-p no:warnings"
